Make go-staticcheck happy plus moving studio commands into own sub-package
